requirements.txt
Flask==2.2.3
transformers==4.30.0
torch==2.0.1
newspaper3k==0.2.8

